fix: точность отображения памяти до сотых + исправление WebSocket автообновления
Изменения:
- DashboardStatsDto: Long → Double для полей памяти (системная ОЗУ и JVM Heap)
- DashboardServiceImpl: вычисление памяти через double деление с Math.round() для точности 2 знака
- index.html:
  * Удалена строка с несуществующим элементом totalFiles (вызывала JavaScript ошибку)
  * Исправлен селектор totalSize (querySelector → getElementById)
  * Добавлен .toFixed(2) для отображения памяти в JavaScript
  * Добавлен try-catch в updateStats() для защиты от ошибок
  * Добавлено Thymeleaf форматирование #numbers.formatDecimal() для начального отображения
  * Перемещен индикатор updateIndicator внутрь секции (исправлена проблема с отображением)

Результат:
- Память отображается с точностью до сотых (15.33 ГБ вместо 15 ГБ)
- WebSocket автообновление работает каждые 15 секунд
- Индикатор "Обновлено" корректно отображается
- Нет JavaScript ошибок в Console

+++ src/main/java/com/java/dto/DashboardStatsDto.java
@@ -82,15 +82,15 @@ public class DashboardStatsDto {
         private String javaVersion;
         private String springBootVersion;
 
-        // JVM Heap память
-        private Long jvmTotalMemoryMb;
-        private Long jvmUsedMemoryMb;
-        private Long jvmFreeMemoryMb;
+        // JVM Heap память (МБ) - с точностью до сотых
+        private Double jvmTotalMemoryMb;
+        private Double jvmUsedMemoryMb;
+        private Double jvmFreeMemoryMb;
 
-        // Системная память ПК (ОЗУ)
-        private Long systemTotalMemoryGb;
-        private Long systemUsedMemoryGb;
-        private Long systemFreeMemoryGb;
+        // Системная память ПК (ОЗУ) (ГБ) - с точностью до сотых
+        private Double systemTotalMemoryGb;
+        private Double systemUsedMemoryGb;
+        private Double systemFreeMemoryGb;
 
         private String operatingSystem;
         private String databaseUrl;
